Handle data that pushes from sensor master

+++ ionoff.server/src/main/java/net/ionoff/center/server/entity/SensorData.java
@@ -19,6 +19,6 @@ public class SensorData implements IEntity {
 	private String name;
 	private Date time;
 	private Double value;
-	private Long index;
+	private Integer index;
 	private Sensor sensor;
 }
